requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements/requirements.in
#
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via litellm
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.6.0
    # via
    #   httpx
    #   openai
attrs==24.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via -r requirements/requirements.in
beautifulsoup4==4.12.3
    # via -r requirements/requirements.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   sounddevice
    #   soundfile
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via litellm
configargparse==1.7
    # via -r requirements/requirements.in
diff-match-patch==20230430
    # via -r requirements/requirements.in
diskcache==5.6.3
    # via -r requirements/requirements.in
distro==1.9.0
    # via openai
filelock==3.16.1
    # via huggingface-hub
flake8==7.1.1
    # via -r requirements/requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.9.0
    # via huggingface-hub
gitdb==4.0.11
GitPython==3.1.43
    # via gitpython
gitpython==3.1.43
    # via -r requirements/requirements.in
grep-ast==0.3.3
    # via -r requirements/requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via openai
huggingface-hub==0.25.0
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==7.2.1
    # via
    #   -r requirements/requirements.in
    #   litellm
importlib-resources==6.4.5
    # via -r requirements/requirements.in
jinja2==3.1.4
    # via litellm
jiter==0.5.0
    # via openai
json5==0.9.25
    # via -r requirements/requirements.in
jsonschema==4.23.0
    # via
    #   -r requirements/requirements.in
    #   litellm
jsonschema-specifications==2023.12.1
    # via jsonschema
litellm==1.47.0
    # via -r requirements/requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
networkx==3.2.1
    # via -r requirements/requirements.in
numpy==1.26.4
    # via
    #   -r requirements/requirements.in
    #   scipy
openai==1.47.0
    # via litellm
packaging==24.1
    # via
    #   -r requirements/requirements.in
    #   huggingface-hub
pathspec==0.12.1
    # via
    #   -r requirements/requirements.in
    #   grep-ast
pexpect==4.9.0
    # via -r requirements/requirements.in
pillow==10.4.0
    # via -r requirements/requirements.in
prompt-toolkit==3.0.47
    # via -r requirements/requirements.in
psutil==6.0.0
    # via -r requirements/requirements.in
ptyprocess==0.7.0
    # via pexpect
pycodestyle==2.12.1
    # via flake8
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   litellm
    #   openai
pydantic-core==2.23.4
    # via pydantic
pydub==0.25.1
    # via -r requirements/requirements.in
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via rich
pypandoc==1.13
    # via -r requirements/requirements.in
pyperclip==1.9.0
    # via -r requirements/requirements.in
python-dotenv==1.0.1
    # via litellm
pyyaml==6.0.2
    # via
    #   -r requirements/requirements.in
    #   huggingface-hub
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.9.11
    # via tiktoken
requests==2.32.3
    # via
    #   huggingface-hub
    #   litellm
    #   tiktoken
rich==13.8.1
    # via -r requirements/requirements.in
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
scipy==1.13.1
    # via -r requirements/requirements.in
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sounddevice==0.5.0
    # via -r requirements/requirements.in
soundfile==0.12.1
    # via -r requirements/requirements.in
soupsieve==2.6
    # via beautifulsoup4
tiktoken==0.7.0
    # via litellm
tokenizers==0.19.1
    # via
    #   -r requirements/requirements.in
    #   litellm
tqdm==4.66.5
    # via
    #   huggingface-hub
    #   openai
tree-sitter==0.21.3
    # via
    #   -r requirements/requirements.in
    #   tree-sitter-languages
tree-sitter-languages==1.10.2
    # via grep-ast
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.3
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
yarl==1.11.1
    # via aiohttp
zipp==3.20.2
    # via importlib-metadata
psycopg2-binary==2.9.9
    # via -r requirements/requirements.in
pgvector==0.2.5
    # via -r requirements/requirements.in
